requirements.txt
streamlit
groq
mcp
httpx-sse # Mantenha esta para garantir a dependência da SseClientTransport
python-dotenv
streamlit-asyncio # Essencial para esta abordagem